Removed plugin id 'kotlin-android-extensions' because it's been deprecated.
Android synthetics is associated with it, and also @Parcelize (I need to migrate @Parcelize to whatever is LTS now).

I have problems in updating entries in the database.

Also, I need to figure out how to use LiveData<> to retrieve data in the Repository when the @Query also requires a parameter to be passed to it.

Looking into Jetpack Compose architecture component...